Fix #1 : Use screenX/screenY for dragend events
On firefox, clientX/pageX are set to 0 on the drag events.
https://bugzilla.mozilla.org/show_bug.cgi?id=505521

A special exception is screenX/screenY which are present on
the dragend event.

clientX/Y could be used by refactoring to handle the "drop" event
instead but LOC for LOC this is simpler and seems to work both on
Chrome and FF. Since "screenX" is the physical screen, I am not sure
if this can have some weird effects, but yolo.