Add pessimistic locking mechanism for items
User actions (selection/edit etc) will try to obtain lock for
the corresponding item(s). Interacting with the items is not
permitted if the lock is held by another user.

Locks have a TTL of 10 seconds that is renewed on each new action.
TTL should hopefully prevent any deadlocks that might otherwise
occur due to races.