only deselect/de-edit/de-drag when someone else is holding lock
Previously app would deselect everything that the user did
not have lock for -- even in the absence of any lock. Changing
so that only if someone else is holding the lock according to
server, will the deselection action be made.

Also minor fixes like uncommenting out dragstart state change
from editing to dragging, renaming lockAndBringToFront as jsut
bringToFront because bringing to front locks the card anyway.